# docs/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "An API to manage movies",
    "version": "1.0.0",
    "title": "Movie Store API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/movies/{id}": {
      "parameters": [
        {
          "name": "id",
          "description": "The id of the movie",
          "in": "path",
          "schema": {
            "type": "string"
          },
          "required": true
        }
      ],
      "get": {
        "tags": ["Movies"],
        "summary": "Get a movies in the store by id",
        "description": "",
        "operationId": "getMovieById",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          }
        }
      },
      "patch": {
        "tags": ["Movies"],
        "summary": "Updates a movie in the store by id",
        "description": "",
        "operationId": "patchMovieById",
        "requestBody": {
          "$ref": "#/components/requestBodies/moviePatchRequest"
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Movies"],
        "summary": "Deletes a movie in the store by id",
        "description": "",
        "operationId": "deleteMovieById",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          }
        }
      }
    }
  },
  "components": {
    "requestBodies": {
...
      "moviePatchRequest": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PartialMovie"
            }
          }
        }
      }
    },
    "responses": {
      "notFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/FailResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "PartialMovie": {
        "type": "object",
        "example": {
          "title": "The Godfather: Part III",
          "year": 2024,
          "imdbRating": 5.4
        },
        "properties": {
          "title": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "length": {
            "type": "number"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "stars": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "director": {
            "type": "string"
          },
          "plot": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          },
          "imdbRating": {
            "type": "number"
          },
          "imdbUrl": {
            "type": "string"
          }
        }
      },
...
      "MovieResponse": {
        "type": "object",
        "required": ["data", "code", "type"],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "default": null,
            "items": {
              "type": "object"
            }
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        },
        "example": {
          "code": 200,
          "type": "Ok",
          "message": null,
          "errors": null,
          "data": {
            "_id": "61fbb717b32b5fef4970fe78",
            "title": "The Godfather: Part II",
            "year": 1974,
            "length": 202,
            "genres": ["Crime", "Drama"],
            "description": "The early life and career of Vito Corleone in 1920s New York City is portrayed, while his son, Michael, expands and tightens his grip on the family crime syndicate.",
            "stars": [
              "Al Pacino",
              "Robert De Niro",
              "Robert Duvall",
              "Diane Keaton"
            ],
            "director": "Francis Ford Coppola",
            "plot": "The continuing saga of the Corleone crime family tells the story of a young Vito Corleone growing up in Sicily and in 1910s New York; and follows Michael Corleone in the 1950s as he attempts to expand the family business into Las Vegas, Hollywood and Cuba.",
            "poster": "https://m.media-amazon.com/images/M/MV5BMWMwMGQzZTItY2JlNC00OWZiLWIyMDctNDk2ZDQ2YjRjMWQ0XkEyXkFqcGdeQXVyNzkwMjQ5NzM@._V1_UY98_CR1,0,67,98_AL_.jpg",
            "imdbRating": 9.1,
            "imdbUrl": "https://www.imdb.com/title/tt0071562/",
            "__v": 0
          }
        }
      },
...
      "FailResponse": {
        "type": "object",
        "required": ["code", "type"],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "default": null
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "default": null
          },
          "data": {
            "type": "object",
            "default": null
          }
        },
        "example": {
          "data": null,
          "code": 400,
          "type": "Bad Request",
          "message": null,
          "errors": null
        }
      }
    }
  }
}